refactor: extract Vulkan buffer and transfer operations into transfer module
Rollback trigger: issues with Vulkan memory operations or test failures
